24_2_26
1. 적 그래플링시 적 위에 바가 생성되며, 바가 가운데에 위치했을 때 키를 누르면 공격이 세게 들어가고, 그래플링 속도 높이기. -2. (24.2_26완)

+++ GrapperStart/Assets/1_Scripts/Enemy/BloodWorker/ScriptableObject/BwData.cs
@@ -6,8 +6,8 @@ using UnityEngine;
 public class BwData : ScriptableObject
 {
 
-    public int bwHp = 10;
-    public void DamagedHp(int amount)
+    public float bwHp = 10.0f;
+    public void DamagedHp(float amount)
     {
         bwHp -= amount;
     }
